+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsejsonschema.py
@@ -40,9 +40,9 @@ class EvaluatorResponseJSONSchema(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-03T09:53:14.266Z"
+    created: Optional[str] = "2026-07-03T12:12:55.302Z"
 
-    updated: Optional[str] = "2026-07-03T09:53:14.266Z"
+    updated: Optional[str] = "2026-07-03T12:12:55.303Z"
 
     updated_by_id: OptionalNullable[str] = UNSET