+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -87,7 +87,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-07-03T09:53:13.066Z")
+    updated: Optional[datetime] = parse_datetime("2026-07-03T12:12:54.165Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")