+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWKPBP7DMHDZC6CTRCJ5H9SN"
+        "01KWKYBAW0E7JM8Y7HEQV24SD5"
     )
     r"""The unique identifier of the data source"""