+++ packages/orq-rc/RELEASES.md
@@ -9259,3 +9259,13 @@ Based on:
 - [python v4.11.0-rc.76] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.76] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.76 - packages/orq-rc
+
+## 2026-07-03 12:15:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.788.0 (2.915.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.77] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.77] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.77 - packages/orq-rc